refactor: Controller를 삭제하고 ChessGame을 Controller로 간주
